+++ lib/src/node/exports.dart
@@ -4,12 +4,15 @@
 
 import 'package:js/js.dart';
 
+import 'types.dart';
+
 @JS()
 class Exports {
   external set run_(function);
   external set render(function);
   external set renderSync(function);
   external set info(String info);
+  external set types(Types types);
 }
 
 @JS()